Extract parsing song XML and getting list of songs to a shared library.
